pyproject.toml
[project]
name = "higuma-dashboard"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Atori Ikeyama", email = "49859718+Atori-Ikeyama@users.noreply.github.com" }
]
dependencies = [
    "streamlit>=1.36.0",
    "folium>=0.17.0",
    "streamlit-folium>=0.21.0",
    "matplotlib>=3.9.1",
    "numpy>=2.0.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
